server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-pagination>=0.12.34",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]
